Update do_create_dft_blocks_lhmodes.m
rho0 vidash vjdash has a nonzero mean, so it is best to subtract this DC component before doing the temporal DFT. This will actually only affect the m = 0, St = 0 component of the Tij spectrum, which is currently very large. However, this will not affect the sound due to the omega^2 term in the Green's function, which means no sound is emitted at St=0.